chore: refactor CloudFileManager into its own module
